deploy v6.5

+++ database/migrations/2022_02_13_172705_create_assets_table.php
@@ -17,7 +17,7 @@ class CreateAssetsTable extends Migration
             $table->bigIncrements('id');
             $table->string('type');
             $table->string('warranty');
-            $table->integer('value');
+            $table->double('value', 8, 2);
             $table->unsignedBigInteger('client_id');
             $table->timestamps();
