+++ database/migrations/2022_02_21_124130_create_families_table.php
@@ -17,7 +17,7 @@ class CreateFamiliesTable extends Migration
             $table->bigIncrements('id');
             $table->integer('amount');
             $table->string('description');
-            $table->integer('value');
+            $table->double('value', 8, 2);
             $table->string('state');
             $table->unsignedBigInteger('client_id');
             $table->timestamps();
